Breaking up logic into separate files
