+++ web/README.md
@@ -1,0 +1,12 @@
+# GoBench Web
+Web interface for GoBench running on Appengine
+
+## Develop
+```
+goapp serve
+```
+
+## Deploy
+```
+go generate
+```
